Alteração do nome das classes.
